Payment transaction changes 1. Add billingGuid to Insert_NewBillingTransaction 22. change siteguid to long

+++ DAL/BillingDAL.cs
@@ -1365,7 +1365,7 @@ namespace DAL
             }
         }
 
-        public static int InsertPaymentGWTransaction(int groupID, long domainId, int siteGuid, PaymentGWTransaction paymentGWTransaction)
+        public static int InsertPaymentGWTransaction(int groupID, long domainId, long siteGuid, PaymentGWTransaction paymentGWTransaction)
         {
             try
             {
